backend/requirements.txt
fastapi
uvicorn
gradio
httpx
SQLAlchemy
asyncpg
python-dotenv
reportlab
pillow
pydantic
python-multipart
Jinja2
MarkupSafe
typing_extensions
annotated-types
anyio
sniffio
h11
httpcore
idna
certifi
charset-normalizer
packaging
six
pytz
tzdata
greenlet
nest_asyncio
filelock
colorama
psutil
aiohttp
bcrypt
python-jose[cryptography]
portalocker